Kept only ci_sanity tests for floatingip tests
